Refactor rcd unit test and move it to test/
