--- litedram/DDR5RCD01/tests/test_top.py
@@ -1,7 +1,0 @@
-#
-# This file is part of LiteDRAM.
-#
-# Copyright (c) 2023 Antmicro <www.antmicro.com>
-# SPDX-License-Identifier: BSD-2-Clause
-
-# Placeholder for the integration test
